requirements-versioned.txt
sentry-sdk==1.15.0
aiocontextvars==0.2.2
PyYAML==6.0
jsonpickle==3.0.1
pyyaml-include==1.3.1
